backend/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Redis
redis>=5.0.0

# HTTP Client
httpx>=0.25.0

# Celery (background tasks)
celery>=5.3.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Excel
pandas>=2.0.0
openpyxl>=3.1.0
